+++ src/components/interaction/loading-indicator/loading-indicator.vue
@@ -1,6 +1,6 @@
 <template>
-	<div class="flex items-center gap-2" aria-live="polite" data-test="loading-indicator">
-		<icon-loading class="spin text-grey-500 dark:text-grey-400" />
+	<div class="flex items-center gap-2" :class="{ 'flex-col': large }" aria-live="polite" data-test="loading-indicator">
+		<icon-loading class="spin text-grey-500 dark:text-grey-400" :class="{ 'size-7': large }" />
 		<slot />
 	</div>
 </template>
@@ -10,4 +10,14 @@
  * `loading-indicator` provides a consistent loading notice and animation when
  * data is pending.
  */
+
+defineProps({
+	/**
+	 * Whether to display a larger version of the loading indicator.
+	 */
+	large: {
+		type: Boolean,
+		default: false,
+	},
+});
 </script>